+++ src/lib/api/GraphQLQueries-fragments.ts
@@ -198,6 +198,7 @@ dcaCondition
     dcaVolumeRequiredChangeRef
     dcaVolumeMaxValue
 skipBalanceCheck
+rejectBelowExchangeMin
 baseSlOn
 closeByTimer
                     closeByTimerValue